using default modal, removed unused component (#1988)
