+++ wondrous-bot-admin/src/components/LevelsReward/index.tsx
@@ -1,24 +1,36 @@
 import { RoundedSecondaryButton } from "components/Shared/styles";
 import AddIcon from "@mui/icons-material/Add";
-import { forwardRef, useContext, useState } from "react";
-import { Box, ButtonBase, ClickAwayListener, Grid, Popper, Typography } from "@mui/material";
-import { useDiscordRoles } from "utils/discord";
+import { useContext } from "react";
+import { Box, ButtonBase } from "@mui/material";
 import GlobalContext from "utils/context/GlobalContext";
-import { Label } from "components/CreateTemplate/styles";
 import { StyledViewQuestResults } from "components/ViewQuestResults/styles";
 import { CloseIcon } from "components/Shared/DatePicker/Icons";
 import { useMutation } from "@apollo/client";
 import { ADD_ORG_LEVEL_REWARD, REMOVE_ORG_LEVEL_REWARD } from "graphql/mutations";
-import { RewardConfigModal } from "components/LevelReward/RewardConfigModal";
-import { LevelsWrapper } from "./styles";
+import RewardModal, { useAddRewardModalState } from "components/CreateTemplate/RewardModal";
+import { PAYMENT_OPTIONS } from "components/CreateTemplate/RewardUtils";
+import { NFTIcon, PoapIcon, TokensIcon } from "components/Icons/Rewards";
+
+interface AddOrgLevelRewardInput {
+  orgId: string;
+  level: number;
+  type: string;
+  discordRewardData?: any;
+  paymentMethodId?: string;
+  paymentMethod?: any;
+  amount?: number;
+  poapRewardData?: any;
+}
 
 const LevelsReward = ({ rewards, discordRoles, level, refetchLevelRewards }) => {
   // FIXME we should pass reward current level rewards probably
   // need to fetch somewhere?
-  const [isRewardModalOpen, setIsRewardModalOpen] = useState(false);
   const allRoles = discordRoles.map((role) => role.roles).flat();
 
+  const rewardModalState = useAddRewardModalState();
+
   const { activeOrg } = useContext(GlobalContext);
+  const { setIsRewardModalOpen } = rewardModalState;
 
   const [removeOrgLevelReward] = useMutation(REMOVE_ORG_LEVEL_REWARD, {
     refetchQueries: ["getOrgLevelsRewards"],
@@ -31,15 +43,46 @@ const LevelsReward = ({ rewards, discordRoles, level, refetchLevelRewards }) => 
     });
     refetchLevelRewards();
   };
+
+  const [addOrgLevelReward] = useMutation(ADD_ORG_LEVEL_REWARD, {
+    refetchQueries: ["getOrgLevelsRewards"],
+  });
+
+  const onRewardAdd = async (reward) => {
+    const isCommunityBadge = reward?.type === PAYMENT_OPTIONS.COMMUNITY_BADGE;
+    let input: AddOrgLevelRewardInput = {
+      orgId: activeOrg?.id,
+      level: level,
+      type: reward?.type,
+    };
+    if (reward?.type === PAYMENT_OPTIONS.TOKEN || isCommunityBadge) {
+      input.paymentMethodId = reward?.paymentMethodId;
+      input.amount = isCommunityBadge ? null : parseInt(reward?.amount);
+      input.type = PAYMENT_OPTIONS.TOKEN;
+    }
+    if (reward?.type === PAYMENT_OPTIONS.DISCORD_ROLE) {
+      input.discordRewardData = reward?.discordRewardData;
+    }
+    if (reward?.type === PAYMENT_OPTIONS.POAP) {
+      input.poapRewardData = reward?.poapRewardData;
+    }
+    addOrgLevelReward({
+      variables: {
+        input,
+      },
+    }).then(() => {
+      refetchLevelRewards();
+    });
+  };
+
   return (
     <>
-      <RewardConfigModal
-        isRewardModalOpen={isRewardModalOpen}
-        setIsRewardModalOpen={setIsRewardModalOpen}
-        level={level}
-        refetchLevelRewards={refetchLevelRewards}
+      <RewardModal
+        handleRewardModalToggle={() => setIsRewardModalOpen(false)}
+        handleOnRewardAdd={onRewardAdd}
+        rewardModalState={rewardModalState}
       />
-      <Box style={{display: 'flex', gap: 5}}>
+      <Box style={{ display: "flex", gap: 5 }}>
         <div style={{ display: "flex", gap: 5 }}>
           {rewards?.length &&
             rewards?.map((reward) => (
@@ -76,7 +119,7 @@ const ExistingLevelsReward = ({ reward, allRoles, level, refetchLevelRewards, ha
 
   return (
     <StyledViewQuestResults>
-      {reward.type === "discord_role" && (
+      {reward.type === PAYMENT_OPTIONS.DISCORD_ROLE && (
         <>
           <img
             src="/images/discord-official-logo.png"
@@ -92,9 +135,18 @@ const ExistingLevelsReward = ({ reward, allRoles, level, refetchLevelRewards, ha
           </ButtonBase>
         </>
       )}
-      {reward.type === "token" && (
+      {reward.type === PAYMENT_OPTIONS.TOKEN && (
         <>
+          {reward?.paymentMethod?.type === PAYMENT_OPTIONS.COMMUNITY_BADGE ? <NFTIcon /> : <TokensIcon />}
           {reward?.amount} {reward?.paymentMethod?.name}
+          <ButtonBase onClick={() => handleRemove()}>
+            <CloseIcon />
+          </ButtonBase>
+        </>
+      )}
+      {reward?.type === PAYMENT_OPTIONS.POAP && (
+        <>
+        {reward?.poapRewardData?.name} <PoapIcon />
           <ButtonBase onClick={() => handleRemove()}>
             <CloseIcon />
           </ButtonBase>
